# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple swagger implementation in Go HTTP",
        "title": "Go Restful API with Swagger",
        "contact": {
            "name": "Linggar Primahastoko",
            "url": "https://linggar.asia",
            "email": "x@linggar.asia"
        },
        "version": "1.0"
    },
    "host": "3117-176-52-104-232.ngrok.io/api",
    "basePath": "/",
    "paths": {
        "/user/entry": {
            "post": {
                "description": "User auth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User auth",
                "operationId": "sign-in",
                "parameters": [
                    {
                        "description": "SignIn params",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SignInParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "User not exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "User list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User list",
                "operationId": "list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "name": "timestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/new": {
            "post": {
                "description": "Create new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "operationId": "sign-up",
                "parameters": [
                    {
                        "description": "SignUp params",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SignUpParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "This username/email is taken",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "description": "User profile data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User profile data",
                "operationId": "profile",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/refresh": {
            "put": {
                "description": "Refresh user token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Refresh user token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upgrade": {
            "get": {
                "description": "Upgrade user role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Upgrade user role",
                "operationId": "upgrade",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user.SignInParams": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456789"
                },
                "username": {
                    "type": "string",
                    "example": "example"
                }
            }
        },
        "user.SignUpParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "123456789"
                },
                "username": {
                    "type": "string",
                    "minLength": 5,
                    "example": "example"
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "basic"
        }
    }
}